+++ Sources/RichTextKit/Actions/RichTextAction.swift
@@ -11,6 +11,9 @@ import SwiftUI
 /**
  This enum defines rich text actions that can be executed on
  a rich text editor.
+ 
+ This type also serves as a type namespace for other related
+ types and views, like ``RichTextAction/Button``.
  */
 public enum RichTextAction: Identifiable, Equatable {